Optimize Sign language

+++ catscripts/Language/Sign_language/Signlanguage.sh
@@ -1,14 +1,30 @@
 #!/bin/bash
 
-egrep -i 'Sign(| )language' newpages.txt > Signlanguage.txt
+KEYWORDS_SIGNLANGUAGE="Sign(| )language"
 
-SIGN=`stat --print=%s Signlanguage.txt`
-
-if [ $SIGN -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="Signlanguage.txt"
-  export CATNAME="Sign language"
-  $CATEGORIZE
-fi
 
-rm Signlanguage.txt
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Starting Sign language\n"
+  fi
+
+  SIGN=`egrep -i "$KEYWORDS_SIGNLANGUAGE" newpages.txt`
+
+  if [ "$SIGN" != "" ];
+  then
+    printf "$SIGN" > Signlanguage.txt
+    export CATFILE="Signlanguage.txt"
+    export CATNAME="Sign language"
+    $CATEGORIZE
+    rm Signlanguage.txt
+    unset SIGN
+  fi
+
+  if [ "$DEBUG" == "yes" ];
+  then
+    printf "Finishing Sign language\n"
+  fi
+
+fi
